Add path_expand and PATTERN token type
We use this new token type in order to add a new child to the token node
for each match if the token node contains a pattern matching expression.
This makes it easy to add each match as a separate argv.
(this commit segfaults on "ls *" because the rest of the code does not
account for the extra children with char * data)