Fix globbing quotes bugs
Using a static to store quote status would not work because of
wildcards, where we recurse twice and quote status remains the same for
both calls, even though they are at different string indices.
We also need to close quotes after we check for matches with the special
globbing characters and we must not advance the string if there is a
quote_stop and the quote was a backslash, in this case the character
after the backslash needs to match exactly.